lanjutan

+++ src/script/view/view.js
@@ -1,5 +1,6 @@
 import ApiNotes from "../data/apiNotes";
-import Swal from "sweetalert2";
+import { showSuccessAlert, showErrorAlert } from "./alerts";
+import showArchivedNotes from "./archiveNotes";
 
 const view = () => {
   const list = document.querySelector(".list");
@@ -7,10 +8,19 @@ const view = () => {
   const formInput = document.querySelector("form-input");
   const titleField = formInput.shadowRoot.querySelector(".input-title");
   const bodyField = formInput.shadowRoot.querySelector("#body-input");
-
+  const customHeader = document.querySelector('custom-header')
+  const buttonOpenDialog = customHeader.shadowRoot.querySelector('.open-dialog')
+  const customDialog = document.querySelector('#archivedNotesDialog')
+  const buttonCloseDialog = customDialog.querySelector('.close-dialog')
   const searchBarField = searchBar.shadowRoot.querySelector(".search__field");
-
   let filteredNotes = [];
+
+  buttonOpenDialog.addEventListener('click', ()=>{
+    customDialog.showModal();
+  })
+  buttonCloseDialog.addEventListener('click', ()=>{
+    customDialog.close();
+  })
 
   formInput.shadowRoot.addEventListener("submit", (e) => {
     e.preventDefault();
@@ -80,48 +90,34 @@ const view = () => {
   };
 
   list.addEventListener("click", (event) => {
-    const clickedButton = event.target.closest(".archive-button");
-    if (clickedButton) {
-      const noteId = clickedButton.dataset.noteId;
-      console.log(noteId);
-      // const success = AllNotes.toggleArchiveNote(noteId);
-      // if (success) {
-      //   showAllNotes();
-      // } else {
-      //   console.error("Failed to toggle archive status for note:", noteId);
-      // }
+    const clickedArchiveButton = event.target.closest(".archive-button");
+    const clickedDeleteButton = event.target.closest(".delete-button");
+  
+    if (clickedArchiveButton) {
+      const noteId = clickedArchiveButton.dataset.noteId;
+      ApiNotes.archiveNote(noteId)
+        .then((message) => {
+          showSuccessAlert(message);
+          showAllNotes();
+          showArchivedNotes()
+        })
+        .catch((error) => {
+          showErrorAlert(error);
+        });
+    }
+  
+    if (clickedDeleteButton) {
+      const noteId = clickedDeleteButton.dataset.noteId;
+      ApiNotes.deleteNote(noteId)
+        .then((message) => {
+          showSuccessAlert(message);
+          showAllNotes();
+        })
+        .catch((error) => {
+          showErrorAlert(error);
+        });
     }
   });
-
-  list.addEventListener("click", (e) => {
-    const clickedButton = e.target.closest(".delete-button");
-    const noteId = clickedButton.dataset.noteId;
-    ApiNotes.deleteNote(noteId)
-      .then((message) => {
-        showSuccessAlert(message);
-        showAllNotes()
-      })
-      .catch((error) => {
-        showErrorAlert(error);
-      });
-  });
-
-  const showSuccessAlert = (message) => {
-    Swal.fire({
-      icon: "success",
-      title: message,
-      showConfirmButton: false,
-      timer: 1500,
-    });
-  };
-  const showErrorAlert = (error) => {
-    Swal.fire({
-      icon: "error",
-      title: error.message,
-      showConfirmButton: false,
-      timer: 1500,
-    });
-  };
 
   showAllNotes();
 };
